Split otel.test.ts into otel-noop.test.ts and otel-sdk.test.ts for proper isolation
Agent-Logs-Url: https://github.com/microsoft/Agents-for-js/sessions/65f3b397-c05c-4756-82cb-06f2ce6fe897
